Split DocumentStateStrategy into separate files.
